Update buildscript for NDKr16b and support for clang
GCC, armv5 and mips are depreciated for the newer versions of NDK and
soon will be removed. This change is able to build the code with either
GCC and clang. armv5 and mips do not build with clang because they have
issues with linking and are depreciated. GCC will receive no more NDK
support and therefore clang will build as default. Also simplified and
streamlined the build process per arch for building newest version of
NDK

As of NDKr16b, armv-7 neon does not build for GCC, please use NDK10e or
something earlier that works. Clang will only build x86 and armv7/neon
and soon 64 bit.

To use GCC, uncomment the at the beginnning of build_android.sh. The
location of standalone toolchain for clang can also be changed.